finish delete trainees

+++ src/main/java/com/thoughtworks/capability/gtb/entrancequiz/repository/TraineeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.repository.CrudRepository;
 
 @Repository
-public interface TraineeRepository extends CrudRepository<Trainee, Integer>{
+public interface TraineeRepository extends CrudRepository<Trainee, Long>{
 
 }
